FooDine
FooDine

Concept: The app is created to serve the food ordering market and can
be a direct competitor to companies like Food Panda.

The final version of the app (which we will present at the end of the
session) intends to fill the following market gaps.

Allow the user to choose from a list of restaurants
Show the list of food categories in any given restaurant (drop down)
Show the list of food items in any given category
Allow the person to place order in quantity of his choice
See the order cart before confirming the order
Shoot a mail or SMS to a given phone number of email ID as an order
confirmation

Each of the above intents will have sub objectives and mile stones.
Since the project is not being imported into Github which is a version
control platform for all open source projects (like the one we are
building), We will add all the further tas ks and milestones on the
same.

Request all of you to install Github for Windows/Linux based on the
preference of your OS and your IDE. In case you are not comfortable
doing it alone, we can do it first thing we meet.

Available Resources:

3 Amateur developer guided by a mentor Mr. Shakti
3 laptops (One with Zaheed and two with me)
Office space for collaborative working
Complete blue print of the project
Sample source codes and android tutorials on all computers

Timelines:
Give a dhinchak name to the project
The first milestone should be complete on or before 6 PM Monday 16 Feb
2015.
The complete project plan to be ready and ratified before end of day 16
Feb 2015